cleanup and further refactoring
